added new migrations

+++ BLL/Services/Flowers/Descriptors/UpdateFlowerDescriptor.cs
@@ -11,6 +11,6 @@ namespace BLL.Services.Flowers.Descriptors
         public int FlowerId { get; set; }
         public string FlowerName { get; set; }
         public int FlowerCount { get; set; }
-        public float FlowerCost { get; set; }
+        public decimal FlowerCost { get; set; }
     }
 }
