+++ DAL/Models/Bouquet.cs
@@ -16,6 +16,6 @@ namespace DAL.Models
         public ICollection<BouquetFlower> BouquetsFlowers { get; set; }
         public ICollection<OrderBouquet> OrderBouquets { get; set; }
 
-        public float Price => BouquetsFlowers?.Sum(bf => bf.Flower.FlowerCost * bf.FlowerCount) ?? 0;
+        public decimal Price => BouquetsFlowers?.Sum(bf => bf.Flower.FlowerCost * bf.FlowerCount) ?? 0;
     }
 }
